requirements.txt
pyrogram==1.2.9
feedparser
apscheduler
psycopg2-binary
sqlalchemy==1.3.23
python-dotenv
TgCrypto
